refactor: rename category.utils to category.service
