+++ __tests__/unit/categories/js/commitlint/commitlint.service.spec.ts
@@ -1,7 +1,7 @@
 import { contextState, fileSystem } from '#__tests__/setup-test-context.ts';
-import { huskyIntegration } from '#src/categories/js/commitlint/commitlint.utils.ts';
+import { huskyIntegration } from '#src/categories/js/commitlint/commitlint.service.ts';
 
-describe('commitlint.utils', () => {
+describe('commitlint.service', () => {
   describe('huskyIntegration', () => {
     it('does not run addHook if husky is not installed', async () => {
       await huskyIntegration();
