+++ src/categories/js/commitlint/commitlint.service.ts
@@ -1,4 +1,4 @@
-import { addHook, isHuskyInstalled } from '#src/categories/js/husky/husky.utils.ts';
+import { addHook, isHuskyInstalled } from '#src/categories/js/husky/husky.service.ts';
 import { CLI_NAME } from '#src/categories/js/commitlint/commitlint.const.ts';
 
 export const huskyIntegration = async () => {
