+++ src/categories/js/lint-staged/config/react-ts.config.ts
@@ -3,7 +3,7 @@ import {
   jestMutation,
   prettierMutation,
   stylelintMutation,
-} from '#src/categories/js/lint-staged/lint-staged.utils.ts';
+} from '#src/categories/js/lint-staged/lint-staged.service.ts';
 import type { Config } from '#src/categories/js/lint-staged/config/config.interface.ts';
 
 export const reactTsConfig: Config = {
